refactor: 삭제할 파일이 없는 경우 아무 동작도 하지 않도록 변경

+++ src/main/java/admin/adminsiteserver/common/aws/infrastructure/S3Uploader.java
@@ -55,6 +55,7 @@ public class S3Uploader {
     }
 
     public void delete(FilePathDto filePathDto) {
+        if (filePathDto == null) return;
         String fileName = filePathDto.getFileUrl().replace(cloudfront, "");
         if (amazonS3.doesObjectExist(bucketName, fileName)) {
             amazonS3.deleteObject(new DeleteObjectRequest(bucketName, fileName));
